# Conflicts:
#	node/cli/src/chain_spec.rs
#	node/runtime/src/lib.rs
#	srml/staking/src/lib.rs
#	srml/staking/src/mock.rs
#	srml/staking/src/tests.rs

+++ node-template/runtime/src/lib.rs
@@ -48,10 +48,10 @@ pub type AccountSignature = sr25519::Signature;
 pub type Hash = primitives::H256;
 
 /// Index of a block number in the chain.
-pub type BlockNumber = u64;
+pub type BlockNumber = u32;
 
 /// Index of an account's extrinsic in the chain.
-pub type Nonce = u64;
+pub type Nonce = u32;
 
 /// Balance type for the node.
 pub type Balance = u128;
